Patch from Alastair M. Robinson <blackfive@fakenhamweb.co.uk> to make the ALPS dual trackpad found on Dell CPx machines work. (For 2.4 kernels.)
In order to get my trackpad working in absolute mode I had to make some
modifications to the driver:

Firstly, initialising the trackpad on this model is more complicated
than other ALPS devices; the usual "magic knock" must be performed for
both the trackpad and the pointing device, and I found that I needed a
device reset first to get reliable initialisation.

Secondly, the pointing stick, for obvious reasons, provides relative
data even when the trackpad is in absolute mode.

Thirdly, the guest mouse in the PS/2 port continues to send 3-byte
packets.