Drop now useless files.
The file "FILES" stated which files contained which content, but the list was
outdated and I dropped many of the files anyway
The "usbhid" stuff could be done better using udev rules and from what I know
(and my knowledge is rather limited) no distribution still uses the hotplug
daemon.
The "usbmouse" script was a simple wrapper for calling the synclient.