Disentangle two-finger tap and two-finger scrolling
When two-finger-scrolling quickly and distinctly, the two-finger tapping gets activated, effectively selecting a portion of the selected text. The synaptics code already has logic to prevent taps from
being registered when the fingers move; this patch simply removes what appears to be an unnecessary assumption that only single-finger clicks needs this functionality.

Signed-off-by: Henrik Rydberg <rydberg@euromail.se>