Optional: Set corner buttons to off by default
The corner button functionality might be good on some touchpads, but it seems there are more
cases where it leads to unexpected behavior. To avoid confusion, having this feature turned off
by default makes more sense. However, there might be strong opinions it, so this patch will have
to be considered optional.

Signed-off-by: Henrik Rydberg <rydberg@euromail.se>