man: Document multifinger-scrolling defaults.
If Protocol event is used, the driver initializes defaults based on the
capabilities reported by the kernel driver. Acceleration and edges are based
on the dimensions reported by the kernel. If the kernel reports multi-finger
detection, two-finger vertical scrolling is enabled, horizontal two-finger
scrolling is disabled and edge scrolling is disabled. If no multi-finger
capabilities are reported, edge scrolling is enabled for both horizontal and
vertical scrolling.