[syndaemon] use Xrecord extension in syndaemon to avoid polling.
This patch prevents the polling of the keyboard state. Instead it uses the
XRecord extension of the Xserver for an event triggered notification of key
events. Of course, there is a fallback to the polling when no XRecord
extension is found. This should finally stop complains of syndaemon
preventing good power saving.

Signed-off-by: Christoph Brill <egore911@egore911.de>