eventcomm: fix return value of event_query_is_touchpad.
"ret" is used both for storing the ioctls return value as well as for the
return of event_query_is_touchpad. If an ioctl fails, ret is -1 and we
return this value as a BOOL. The caller treats this -1 as TRUE and thinks
the device is a touchpad.

Bug is triggered if a xorg.conf section is present with auto-dev as device
and a mouse event device is openeded before the synaptics event device.

Introduced with e4b1571d487cb67bab64e1ee890bddcd02437ddf.

Red Hat Bug 499792 <http://bugzilla.redhat.com/show_bug.cgi?id=499792>

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Henrik Rydberg <rydberg@euromail.se>