Revert "Auto-adjust edges if values fall outside queried min/max ranges. (#21001)"
This reverts commit afb60a0b2497c5d08cbd1739fa8ae6585c428881.

From comment 24 to #21001:
I've been running this code for over a week now and I'm not happy with it.
Once I move over to the right, the scroll-edge becomes so small that it's
hard to trigger.

Source of the problem is the information provided by the kernel. The kernel
hands us a min/max value for the synaptics pads but this value is not
reflective of the actual physical boundaries. The other dimensions are based
on these min/max ranges.

My RightEdge setting by default is 5129, after moving to the right it
becomes 5677. The announced max for x is 5472.  We have model-specific edge
settings and in the case of synaptics the width of the scroll area is 7% of
the total width (based on min/max). This works, but obviously only because
the max is wrong. I've tried upping this to 15% and it works fine but unless
the edge is adjusted the scroll bar takes over too much of the pad.

So right now I'm inclined to revert this patch and just ditch any
auto-adjustment of scroll edges whatsoever. This way, the original setting
is maintained even if we reach outside of the min/max area.

Conflicts:

	src/synaptics.c

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>