Add support for reducing the build process output by using shave
This adds suport for shave (git://git.lespiau.name/shave) which is
capable of reducing the amount of unintresting output from autotools
builds. It was copied over from xf86-video-ati and updated to latest git
of shave. It is still possible to get the usual output by using "make
V=1".