Reshuffle default value initialization.
No functional changes, just code structure. The eventcomm backened is the
only one with a real probe function right now. So provide a default query
proc that reads the touchpad coordinates in. The values provided are the
ones that then will give us the right edge settings in concordance with the
default we had before.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>