synaptics: detect corner tap using the initial touch position
Currently, when detecting a tap release, the corner tap detection
uses the current touch position to figure out which corner it
corresponds. The problem is that, theoretically, there is no such position
because there is no touch. It work in mosts cases because most touchpad
are fast enough to still have a position on release, but some aren't.

In that case, the driver do corner tap detection using the position
given by the hardware when there is no touch, that is, (0,0), (1,5855)
or whatever. The driver will detect them as corner tap, regardless of
the initial touch position. On the default configuration, corner taps
are ignored, so basically, these tap are not working.

This patch make tap detection use the initial touch position to
detect corner taps.

Signed-off-by: Nicolas Cavallari <batchman@free.fr>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>