Don't autoprobe for devices when Option Device is set.
If only Option Device is set but no protocol, the code calls into
AutoDevProbe. eventcomm (the only backend with an AutoDevProbe) then runs
through all /dev/input/event devices and takes the first one it can find.

If two touchpads are connected on a system, this may cause the same touchpad
to be added twice and the other one not at all - even though the device path
is specified. (This can only happen when the event device is not grabbed,
otherwise the grabcheck prevents the touchpad from being added twice)

Pass the device option into AutoDevProbe and check that device first. If it
is a touchpad, finish with success. If it isn't, fail AutoDevProbe.

Introduced in dce6006f6a851be4147e16731caa453dd0d1ec1c.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
CC: Alexandr Shadchin <alexandr.shadchin@gmail.com>
Reviewed-by: Chris Bagwell <chris@cnpbagwell.com>