Revert "build: sort building of tools, ensure that cross-pkg-config works."
This reverts commit 4005df66072ceac175ea71427deb16176262f197.

The patch introduces a couple of issues.
1) These tools are apps and conceptually do not depend on Xserver,
so XORG_CFLAGS should not be used. It included pixman header files.
Only drivers depend xserver

2) XORG_CFLAGS may contain, depending on the platform, a compiler
visibilty flag. Compiler flags cannot be assigned to AM_CPPFLAGS
preprocessor flags variable.

There were two changes introduced by the patch.
1) Possible wrong order of include directives
The commit text does not mention which paths may be in the wrong order.

2) Incorrect usage of AM_LDFLAGS
The patch does not really change anything. If the intention was for the tools
to link directly to the libraries, it would look like:

  LDADD =  $(top_builddir)/src/libpciaccess.la # scanpci example

This is only possible for libraries built in the same package.
There is nothing wrong in putting -l -L flags in LDADD, but should there
be other flags, LDFLAGS would be required, ending with two variables.

It looks easier to revert the patch to the previosuly known good version
which has been in service for a year.

Unfortunatly, the reversal breaks distcheck as it removes an unsuspected workaround.
The value of sdkdir during distcheck points to an empty xserver sdkdir
where xserver-properties.h is expected by the tools.

Specifying XORG_CFLAGS worked around the issue by supplying the real path
to the xserver sdk.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>