Add distcheck support for configuration files when dir is not writable
Provide a user writable location for configuration files.
Many values will work, but preserving the semantic by using
a value based on the server default value is more helpful.

The configdir automake variable and the pkgconfig sysconfigdir variable
should not be confused with the sysconfdir automake provided
configuration option which default value is $prefix/etc.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>