Don't emit touch sequences if only one touch is active
When a second touch begins, emit a touch begin for the first touch with
the current valuator values. When a touch ends and we are going from two
touches down to one touch, end both touches. This ensures we don't send
a touch sequence at the same time we are moving the pointer.

Signed-off-by: Chase Douglas <chase.douglas@canonical.com>