Revert "Don't store fake events in the motion history"
This commit introduced a regression. On some touchpads, the pointer keeps
moving in the last direction when the finger movement stops but the finger
is left on the touchpad.

Cause appears to be get_delta() which calculates the deltas based on the
motion history but has no control flow for the lack of fake motion events
in the history after this commit. Thus, under some conditions, the delta is
always non-zero as the history does not change.

Reproducer attached to bug
https://bugs.freedesktop.org/show_bug.cgi?id=45278#c11

X.Org Bug 45278 <http://bugs.freedesktop.org/show_bug.cgi?id=45278>

This reverts commit c8b098214b44cf0585d78c460401ea7d143769f3.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by: Chase Douglas <chase.douglas@canonical.com>
Acked-by: Daniel Stone <daniel@fooishbar.org>