Fix clickfinger actions when middle button emulation is enabled
When MBE is enabled, a physical left button press is delayed until a
timeout is reached. This results in the logical left button being
depressed while the physical left button is pressed. The physical state
is stored as the "old" hw state, and it is used for detecting a
transition from depressed to pressed for clickfinger actions. Since the
"old" hw state shows the left button pressed, but the current logical
state shows the left button unpressed, when the MBE timeout fires and we
set the logical left button pressed the transition check fails.

Since the "old" hw state is only used for clickfinger left button press
transitions, redefining it to hold the previous logical hw state is
sufficient for fixing the bug and should not cause any regressions.

Signed-off-by: Chase Douglas <chase.douglas@canonical.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>