Allow dead areas to update the button state
On clickpads, a dead area helps prevent pointer movement when the buttons
are pressed. Thus check for the button state before resetting the hw state
in case we get a right-click in the soft button area.

For clickfinger, the effect is that clickfinger events are now triggered if
the finger rests in the dead area.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>