Drop the half-baked test suite
This suite was never really maintained anyway, and it is quite hard to do so
anyway. The driver is linked to the server's API too tightly to easily do
independent testing. We need to re-implement stubs for the API the driver
uses, have to track API changes, etc. Not worth the effort.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>